+++ bindings/c/tree-sitter-wdl.h
@@ -1,0 +1,16 @@
+#ifndef TREE_SITTER_WDL_H_
+#define TREE_SITTER_WDL_H_
+
+typedef struct TSLanguage TSLanguage;
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+const TSLanguage *tree_sitter_wdl(void);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif // TREE_SITTER_WDL_H_